requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.55.3 
aiohttp==3.9.1